+++ cpp/src/arrow/sparse_tensor_test.cc
@@ -216,7 +216,7 @@ TYPED_TEST_P(TestSparseCOOTensorForIndexValueType, CreationWithRowMajorIndex) {
   std::vector<c_index_value_type> coords_values = {0, 0, 0, 0, 0, 2, 0, 1, 1, 0, 1, 3,
                                                    0, 2, 0, 0, 2, 2, 1, 0, 1, 1, 0, 3,
                                                    1, 1, 0, 1, 1, 2, 1, 2, 1, 1, 2, 3};
-  const size_t sizeof_index_value = sizeof(c_index_value_type);
+  const int sizeof_index_value = sizeof(c_index_value_type);
   auto si = this->MakeSparseCOOIndex(
       {12, 3}, {sizeof_index_value * 3, sizeof_index_value}, coords_values);
 
@@ -243,7 +243,7 @@ TYPED_TEST_P(TestSparseCOOTensorForIndexValueType, CreationWithColumnMajorIndex)
   std::vector<c_index_value_type> coords_values = {0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
                                                    0, 0, 1, 1, 2, 2, 0, 0, 1, 1, 2, 2,
                                                    0, 2, 1, 3, 0, 2, 1, 3, 0, 2, 1, 3};
-  const size_t sizeof_index_value = sizeof(c_index_value_type);
+  const int sizeof_index_value = sizeof(c_index_value_type);
   auto si = this->MakeSparseCOOIndex(
       {12, 3}, {sizeof_index_value, sizeof_index_value * 12}, coords_values);
 
@@ -271,7 +271,7 @@ TYPED_TEST_P(TestSparseCOOTensorForIndexValueType,
 
   // Row-major COO index
   const std::vector<int64_t> coords_shape = {12, 3};
-  const size_t sizeof_index_value = sizeof(c_index_value_type);
+  const int sizeof_index_value = sizeof(c_index_value_type);
   std::vector<c_index_value_type> coords_values_row_major = {
       0, 0, 0, 0, 0, 2, 0, 1, 1, 0, 1, 3, 0, 2, 0, 0, 2, 2,
       1, 0, 1, 1, 0, 3, 1, 1, 0, 1, 1, 2, 1, 2, 1, 1, 2, 3};
